Extract method client macro factory
